Refactor chat components and implement usage limit management
- Removed the `useUsageLimit` hook and integrated its functionality into the new `useChat` hook.
- Created `ChatClient`, `ChatInput`, and `ChatSidePanel` components to manage chat interactions.
- Implemented auto-scrolling for chat messages using `useAutoScroll` hook.
- Added usage limit checks before sending messages and displaying appropriate warnings.
- Enhanced chat title editing functionality with save and cancel options.
- Improved user experience with responsive design adjustments for mobile and desktop views.
- Introduced utility functions for managing chat input state and validation.
