+++ app/private/page.tsx
@@ -1,5 +1,5 @@
 import { createClient } from "@/utils/supabase/server";
-import ChatClient from "../../components/private/ChatClient";
+import ChatClient from "../../components/private/shared/ChatClient";
 import { redirect } from "next/navigation";
 
 export default async function DashboardPage() {
