+++ components/private/shared/ChatInput.tsx
