+++ hooks/private/mainPanel/PMP/useQuestionFlow.ts
@@ -1,7 +1,7 @@
 // hooks/private/mainPanel/useQuestionFlow.ts
 import type { Message, Phase, QuestionData, Provider } from "@/lib/types";
 import { useQuestionGenerator } from "@/hooks/private/mainPanel/PMP/useQuestionGenerator";
-import { ChatSidePanelHandle } from "@/components/private/ChatSidePanel";
+import { ChatSidePanelHandle } from "@/components/private/shared/ChatSidePanel";
 import { addTypingMessage } from "@/lib/messageHelpers";
 
 interface QuestionFlowLogicProps {
